extracted QtCreator OSD to own ahk script
The QtCreator OSD code was originally in the os-global.ahk, but
it caused problems, the OnWmPowerBroadcast() handler
for the OnMessage(0x218, "OnWmPowerBroadcast") was called twice!
I think it was because the CreateQtCreatorOSD() creates a window behind
the scene, so I got two windows and so it was called twice.

Because of that I have made standalone script
for the CreateQtCreatorOSD().